reshape tentacles and move pulse to agent tentacle
