+++ octomate/memory/zep.py
@@ -14,7 +14,7 @@ from octomate.schemas.events import MessageEvent
 from octomate.schemas.session import SessionKey, UserProfile
 
 if TYPE_CHECKING:
-    from octomate.tentacles.base import ChannelTentacle
+    from octomate.tentacles.channel.base import ChannelTentacle
 
 logger = logging.getLogger(__name__)
