+++ octomate/tentacles/__init__.py
@@ -1,10 +1,6 @@
-from octomate.tentacles.base import (
-    AgentTentacle,
-    ChannelTentacle,
-    Ink,
-    MessageBuffer,
-    Tentacle,
-)
+from octomate.tentacles.agent.base import AgentTentacle
+from octomate.tentacles.base import Tentacle
+from octomate.tentacles.channel.base import ChannelTentacle, Ink, MessageBuffer
 
 __all__ = [
     "AgentTentacle",
