+++ octomate/tentacles/agent/__init__.py
@@ -1,0 +1,19 @@
+from octomate.tentacles.agent.base import AgentTentacle
+from octomate.tentacles.agent.context import SessionContext, SummonRequest
+from octomate.tentacles.agent.pulse import (
+    PulseTentacle,
+    build_pulse_tentacle,
+    build_summon_toolset,
+)
+from octomate.tentacles.agent.skills import SkillInfo, SkillManager
+
+__all__ = [
+    "AgentTentacle",
+    "PulseTentacle",
+    "SessionContext",
+    "SkillInfo",
+    "SkillManager",
+    "SummonRequest",
+    "build_pulse_tentacle",
+    "build_summon_toolset",
+]
