+++ octomate/tentacles/agent/claude/__init__.py
@@ -1,0 +1,3 @@
+from octomate.tentacles.agent.claude.agent import ClaudeCodeTentacle
+
+__all__ = ["ClaudeCodeTentacle"]
