+++ octomate/tentacles/agent/pulse/__init__.py
@@ -1,0 +1,21 @@
+from octomate.tentacles.agent.pulse.graph import ExecuteStep, Synthesize, Triage, pulse_graph
+from octomate.tentacles.agent.pulse.run import resolve_deferred, run_streaming
+from octomate.tentacles.agent.pulse.state import LocalSubAgent, PulseDeps, PulseState, SubAgent, TriageOutput
+from octomate.tentacles.agent.pulse.tentacle import PulseTentacle, build_pulse_tentacle, build_summon_toolset
+
+__all__ = [
+    "ExecuteStep",
+    "LocalSubAgent",
+    "PulseDeps",
+    "PulseState",
+    "PulseTentacle",
+    "SubAgent",
+    "Synthesize",
+    "TriageOutput",
+    "Triage",
+    "build_pulse_tentacle",
+    "build_summon_toolset",
+    "pulse_graph",
+    "resolve_deferred",
+    "run_streaming",
+]
