+++ octomate/tentacles/base.py
@@ -1,118 +1,16 @@
 from __future__ import annotations
 
-import asyncio
 import logging
 from abc import ABC, abstractmethod
-from collections import defaultdict
-from collections.abc import AsyncIterator, Awaitable, Callable
-from contextlib import asynccontextmanager
-from dataclasses import dataclass, field
-from functools import cached_property
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar, Protocol, runtime_checkable
+from typing import TYPE_CHECKING
 
-import anyio
-from pydantic_ai import (
-    CallDeferred,
-    RunContext,
-)
-from pydantic_ai.toolsets import FunctionToolset
-
-from octomate.agents.base import SessionContext
-from octomate.agents.tools import history_toolset
-from octomate.nerve import AgentResult, DismissPending
 from octomate.schemas.events import MessageEvent
-from octomate.schemas.segments import (
-    AgentSegment,
-    ImageSegment,
-    MarkdownSegment,
-    ReplySegment,
-    TextSegment,
-)
-from octomate.schemas.session import SessionKey, UserProfile
-from octomate.stores.interaction import InteractionStore
-from octomate.stores.thread import ThreadStore
-from octomate.tentacles.feelers import NULL_FEELERS, Feelers
+from octomate.schemas.session import SessionKey
 
-# agents.pulse imports from this module (via SessionContext),
-# and octopus imports Tentacle from this module — both circular.
 if TYPE_CHECKING:
-    from octomate.agents.pulse import PulseAgents
-    from octomate.memory.base import OctopusMemory
     from octomate.octopus import Octopus
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class PlatformMessage:
-    msg_type: str
-    content: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@runtime_checkable
-class Chromo(Protocol):
-    """Two-way translation between platform-native wire format and internal schema."""
-
-    async def sip(self, raw: Any) -> MessageEvent | None: ...
-
-    async def squirt(
-        self, segments: list[AgentSegment], *, reply_to: str | None = None
-    ) -> list[PlatformMessage]: ...
-
-
-@runtime_checkable
-class Ink(Protocol):
-    """Structural protocol for platform API clients — identity and media only."""
-
-    def inspect(self) -> UserProfile: ...
-
-    async def get_user_profile(self, user_id: str) -> UserProfile: ...
-
-    async def upload_media(self, data: bytes) -> str | None:
-        """Upload media bytes, return a platform key/URL or None on failure."""
-        ...
-
-    async def download_media(
-        self, resource_id: str, **kwargs: Any
-    ) -> tuple[bytes, str] | None:
-        """Download media by platform resource ID. Returns (data, filename) or None."""
-        ...
-
-
-class StreamSink:
-    """Streaming output sink for agent tentacles.
-
-    Channels provide platform-specific implementations via open_stream().
-    """
-
-    """Default StreamSink that sends each append/status as a separate message."""
-
-    tentacle: ChannelTentacle
-    key: SessionKey
-
-    def __init__(self, tentacle: ChannelTentacle, key: SessionKey) -> None:
-        self.tentacle = tentacle
-        self.key = key
-
-    async def set_status(self, status: str) -> None:
-        await self.tentacle.twitch(self.key, [TextSegment(data={"text": status})])
-
-    async def append(self, text: str) -> None:
-        await self.tentacle.twitch(self.key, [MarkdownSegment(data={"text": text})])
-
-    async def flush(self) -> None:
-        pass
-
-    async def post_thinking_block(self, text: str) -> None:
-        """Post a collapsible thinking block.
-
-        Default: appends the raw text so platforms without folding support
-        still surface the content inline.
-        """
-        if text:
-            await self.append(text)
 
 
 class Tentacle(ABC):
@@ -131,389 +29,3 @@ class Tentacle(ABC):
         key: SessionKey,
         contents: list[MessageEvent],
     ): ...
-
-
-class AgentTentacle(Tentacle):
-    """A tentacle wrapping an agent. On-demand, not long-running.
-    Borrows the calling ChannelTentacle's feelers and ink for user interaction.
-
-    Only one run per session key at a time. If a new message arrives while the
-    agent is running, the current run is cancelled and re-started with the new
-    message. Pending tool calls, questions, and todos from the previous run are
-    dismissed in parallel.
-    """
-
-    description: str = ""
-    handover: bool = False
-    _running_tasks: dict[SessionKey, asyncio.Task]
-
-    def __init__(
-        self,
-        id: str,
-        octopus: Octopus,
-        description: str = "",
-        *,
-        handover: bool = False,
-    ) -> None:
-        super().__init__(id, octopus)
-        self.description = description
-        self.handover = handover
-        self._running_tasks = {}
-
-    async def __call__(
-        self,
-        key: SessionKey,
-        contents: list[MessageEvent],
-        *,
-        session_name: str = "",
-        request_id: str = "",
-        silent: bool = False,
-    ) -> None:
-        existing = self._running_tasks.get(key)
-        if existing is not None and not existing.done():
-            await self.interrupt(key)
-            existing.cancel()
-            asyncio.create_task(self._dismiss_pending(key))
-
-        self._running_tasks[key] = asyncio.current_task()  # type: ignore[assignment]
-        result_output = ""
-        try:
-            result_output = (
-                await self.run(key, contents, session_name=session_name, silent=silent)
-                or ""
-            )
-        except asyncio.CancelledError:
-            logger.info("AgentTentacle %s: run cancelled for [%s]", self.id, key)
-            raise
-        except Exception:
-            logger.exception("Error in agent tentacle %s [%s]", self.id, key)
-            result_output = "Agent encountered an error."
-        finally:
-            if self._running_tasks.get(key) is asyncio.current_task():
-                self._running_tasks.pop(key, None)
-            if request_id:
-                await self.octopus.agent_nerve.send(
-                    AgentResult(key=key, request_id=request_id, output=result_output)
-                )
-
-    async def interrupt(self, key: SessionKey) -> None:
-        """Send a graceful stop signal before hard-cancelling.
-
-        Override in subclasses to e.g. send ClaudeSDKClient.interrupt().
-        """
-
-    @abstractmethod
-    async def run(
-        self,
-        key: SessionKey,
-        contents: list[MessageEvent],
-        *,
-        session_name: str = "",
-        silent: bool = False,
-    ) -> str | None: ...
-
-    async def _dismiss_pending(self, key: SessionKey) -> None:
-        """Background task: expire all pending interactions and dismiss their cards."""
-        try:
-            await self.octopus.agent_nerve.send(DismissPending(key=key))
-        except Exception:
-            logger.exception(
-                "AgentTentacle %s: error sending DismissPending for [%s]", self.id, key
-            )
-
-
-class ChannelTentacle(Tentacle):
-    """A tentacle wrapping a single IM platform connection. It receives events
-    from the IM, resolves media, pushes events into the Nerve, and listens
-    for outbound actions to send back."""
-
-    FILES_ROOT: ClassVar[Path] = Path(".octomate/files")
-
-    profile: UserProfile
-    ink: Ink
-    chromo: Chromo
-    feelers: Feelers
-    agents: PulseAgents
-    memory: OctopusMemory
-    buffer: MessageBuffer
-    user_profiles: dict[str, UserProfile]
-    threads: ThreadStore
-    interactions: InteractionStore
-
-    def __init__(
-        self,
-        id: str,
-        octopus: Octopus,
-        agents: PulseAgents,
-        memory: OctopusMemory,
-        flush_delay: float = 0.5,
-    ) -> None:
-        super().__init__(id, octopus)
-        self.agents = agents
-        self.memory = memory
-        self.buffer = MessageBuffer(flush_delay=flush_delay, handler=octopus.kick)
-        self.user_profiles = {}
-        self.threads = ThreadStore(default_owner=id)
-        self.interactions = InteractionStore(threads=self.threads)
-        self.feelers = NULL_FEELERS
-
-        self.profile = self.ink.inspect()
-        logger.info(
-            "Tentacle %s: probed as %s (%s)",
-            self.id,
-            self.profile.user_id,
-            self.profile.name,
-        )
-
-    @property
-    def name(self) -> str:
-        return self.profile.name
-
-    @abstractmethod
-    async def activate(self) -> None: ...
-
-    @abstractmethod
-    async def deactivate(self) -> None: ...
-
-    async def __call__(self, key: SessionKey, contents: list[MessageEvent]) -> None:
-        """Process an inbound message batch via the pulse graph."""
-        if not contents:
-            return
-        try:
-            from octomate.agents.pulse import PulseDeps, PulseState, Triage, pulse_graph
-
-            await self.memory.record(key, contents)
-
-            if key.group_id and not any(
-                msg.is_at(self.profile.user_id) for msg in contents
-            ):
-                await self.memory.memo(key, contents, self)
-                return
-
-            context = f"[group: {key.group_id}]" if key.group_id else "[chat: private]"
-            header = f"[me: {self.name} ({self.profile.user_id})] {context}"
-            user_prompt: list = [header]
-            for msg in contents:
-                user_prompt.extend(msg.to_content_parts())
-
-            memories = await self.memory.recall(key, contents, self)
-            instructions: str | None = None
-            if memories:
-                facts = "\n".join(f"- {m}" for m in memories)
-                instructions = f"[relevant memories]\n{facts}"
-
-            logger.info(
-                "Tentacle %s pulse [%s] (%d messages)", self.id, key, len(contents)
-            )
-
-            deps = SessionContext(session_key=key, tentacle=self)
-            state = PulseState(prompt=user_prompt)
-            async with self.open_stream(key) as stream:
-                pulse_deps = PulseDeps(
-                    triage=self.agents.triage,
-                    step=self.agents.step,
-                    synthesize=self.agents.synthesize,
-                    agent_deps=deps,
-                    tentacle=self,
-                    toolsets=self.toolsets,
-                    instructions=instructions,
-                    message_history=await self.memory.history(key, size=32),
-                    stream=stream,
-                )
-                graph_result = await pulse_graph.run(
-                    Triage(), state=state, deps=pulse_deps
-                )
-
-                for msg_out in graph_result.output:
-                    await self.twitch(key, msg_out.segments)
-            asyncio.create_task(self.memory.memo(key, graph_result.output, self))
-        except Exception:
-            logger.exception("Error in tentacle %s [%s]", self.id, key)
-
-    async def ingest(self, raw: Any) -> None:
-        """Inbound pipeline: decode → enrich sender → resolve media → triage."""
-        try:
-            event = await self.chromo.sip(raw)
-            if event is None:
-                return
-            event.tentacle_id = self.id
-            event.self_id = self.profile.user_id
-            event.sender = await self.get_user_profile(event.user_id)
-            await self.submerge(event)
-            self.buffer.push(event)
-        except Exception:
-            logger.exception("Tentacle %s: error in ingest", self.id)
-
-    async def twitch(self, key: SessionKey, segments: list[AgentSegment]) -> None:
-        """Outbound pipeline: resolve media → encode → send."""
-        await self.emerge(segments)
-        reply_to: str | None = str(key.thread_id) if key.thread_id else None
-        for seg in segments:
-            if isinstance(seg, ReplySegment):
-                reply_to = seg.data["id"]
-                break
-        remaining: list[AgentSegment] = [
-            s for s in segments if not isinstance(s, ReplySegment)
-        ]
-        messages: list[PlatformMessage] = await self.chromo.squirt(remaining)
-        chat_id = key.chat_id or key.group_id or key.user_id
-        chat_type = "group" if key.group_id else "private"
-        await self.send_platform_message(
-            str(chat_id),
-            chat_type,
-            messages,
-            reply_to,
-        )
-
-    @abstractmethod
-    async def send_platform_message(
-        self,
-        chat_id: str,
-        chat_type: str,
-        messages: list[PlatformMessage],
-        reply_to: str | None = None,
-        reply_in_thread: bool = False,
-    ) -> str | None: ...
-
-    @asynccontextmanager
-    async def open_stream(self, key: SessionKey) -> AsyncIterator[StreamSink]:
-        """Open a streaming sink for progressive output.
-
-        Subclasses override to provide platform-specific streaming (e.g. Slack
-        message updates). The default sends each append/status as a separate message.
-        """
-        sink = StreamSink(tentacle=self, key=key)
-        try:
-            yield sink
-        finally:
-            await sink.flush()
-
-    async def get_user_profile(self, user_id: str) -> UserProfile:
-        cached = self.user_profiles.get(user_id)
-        if cached is not None:
-            return cached
-        profile = await self.ink.get_user_profile(user_id)
-        self.user_profiles[user_id] = profile
-        return profile
-
-    async def submerge(self, event: MessageEvent) -> None:
-        """Resolve inbound media: download images from the event to local storage."""
-        pending = [seg for seg in event.segments if isinstance(seg, ImageSegment)]
-        if not pending:
-            return
-        save = self.den(event)
-        message_id = str(event.message_id)
-        await anyio.Path(save).mkdir(parents=True, exist_ok=True)
-        async with asyncio.TaskGroup() as tg:
-            for seg in pending:
-                tg.create_task(self.absorb(seg, save, message_id))
-
-    async def emerge(self, segments: list[AgentSegment]) -> None:
-        """Prepare outbound media: upload images before sending."""
-        for seg in segments:
-            if isinstance(seg, ImageSegment):
-                await self.secrete(seg)
-
-    @abstractmethod
-    async def absorb(self, seg: ImageSegment, save_dir: Path, message_id: str) -> None:
-        """Download a single inbound image to save_dir."""
-        ...
-
-    @abstractmethod
-    async def secrete(self, seg: ImageSegment) -> None:
-        """Prepare a single outbound image for sending."""
-        ...
-
-    def den(self, event: MessageEvent) -> Path:
-        subdir = event.chat_id if event.chat_type == "group" else event.user_id
-        return self.FILES_ROOT / self.id / subdir
-
-    @cached_property
-    def toolsets(self) -> list[FunctionToolset[SessionContext]]:
-        toolset = FunctionToolset[SessionContext]()
-
-        @toolset.tool
-        async def acknowledge(ctx: RunContext[SessionContext], text: str) -> str:
-            """
-            Send a short message to the user immediately before doing heavy work or
-            provide extra info ahead of next steps to inform the user what is going on.
-
-            Call this FIRST when about to invoke a skill or tool that may take a few
-            seconds (e.g. weather, search, knowledge base), so the user knows you are
-            working on it. Do NOT use for simple replies like greetings or responses
-            that don't involve tool calls. Example: acknowledge("let me look that up~")
-            """
-            if ctx.deps.tentacle:
-                # Pulse always replies in the main thread — sub-thread messages
-                # are routed directly to agent tentacles and never reach here.
-                await ctx.deps.tentacle.twitch(
-                    ctx.deps.session_key, [TextSegment(data={"text": text})]
-                )
-            return "acknowledged"
-
-        @toolset.tool
-        async def ask_user(
-            ctx: RunContext[SessionContext],
-            question: str,
-            options: list[str] | None = None,
-        ) -> str:
-            """Ask the user a question and wait for their answer before continuing.
-
-            ALWAYS USE this tool when you need clarification or a decision from the user.
-            DO NOT SEND a separate text message asking the same thing — this tool handles it.
-            Provide options to show choice buttons; omit for free-text input
-            (platform-dependent — may not be supported everywhere).
-            Returns the user's answer, or '(no response)' on timeout.
-            """
-            raise CallDeferred()
-
-        @toolset.tool
-        async def create_todo(ctx: RunContext[SessionContext], title: str) -> str:
-            """Create a TODO card for the user in the current chat.
-
-            Use this whenever a task has multiple stages or steps — create a todo item
-            for each stage so the user can track progress. Returns a todo ID on success,
-            or an error message if not supported on this platform.
-            """
-            if not ctx.deps.tentacle:
-                return "not supported"
-            key = ctx.deps.session_key
-            item = await ctx.deps.tentacle.feelers.todos.create_todo(key, title)
-            return f"todo:{item.todo_id}" if item else "not supported on this platform"
-
-        return [toolset, history_toolset()]
-
-
-class MessageBuffer:
-    _flush_delay: float
-    _handler: Callable[[SessionKey, list[MessageEvent]], Awaitable[None]]
-    _buckets: defaultdict[SessionKey, list[MessageEvent]]
-    _pending: set[SessionKey]
-
-    def __init__(
-        self,
-        flush_delay: float,
-        handler: Callable[[SessionKey, list[MessageEvent]], Awaitable[None]],
-    ) -> None:
-        self._flush_delay = flush_delay
-        self._handler = handler
-        self._buckets: defaultdict[SessionKey, list[MessageEvent]] = defaultdict(list)
-        self._pending: set[SessionKey] = set()
-
-    def push(self, event: MessageEvent) -> None:
-        key = event.session_key
-        self._buckets[key].append(event)
-        if key not in self._pending:
-            self._pending.add(key)
-            asyncio.create_task(self._flush_after_delay(key))
-
-    async def _flush_after_delay(self, key: SessionKey) -> None:
-        await asyncio.sleep(self._flush_delay)
-        self._pending.discard(key)
-        batch = self._buckets.pop(key, [])
-        if not batch:
-            return
-        try:
-            await self._handler(key, batch)
-        except Exception:
-            logger.exception("Error handling batch for %s", key)
