+++ octomate/tentacles/channel/__init__.py
@@ -1,0 +1,20 @@
+from octomate.tentacles.channel.base import (
+    ChannelTentacle,
+    Chromo,
+    Ink,
+    MessageBuffer,
+    PlatformMessage,
+    StreamSink,
+)
+from octomate.tentacles.channel.feelers import Feelers, NULL_FEELERS
+
+__all__ = [
+    "ChannelTentacle",
+    "Chromo",
+    "Feelers",
+    "Ink",
+    "MessageBuffer",
+    "NULL_FEELERS",
+    "PlatformMessage",
+    "StreamSink",
+]
