+++ octomate/tentacles/channel/lark/__init__.py
@@ -1,0 +1,8 @@
+from octomate.tentacles.channel.lark.base import LarkTentacle
+from octomate.tentacles.channel.lark.ink import LarkInk, LarkUserProfile
+
+__all__ = [
+    "LarkInk",
+    "LarkTentacle",
+    "LarkUserProfile",
+]
