+++ octomate/tentacles/channel/lark/base.py
@@ -19,23 +19,23 @@ from pydantic import SecretStr
 from uuid_utils import uuid7
 
 from octomate.schemas.segments import ImageSegment
-from octomate.tentacles.base import ChannelTentacle, PlatformMessage
-from octomate.tentacles.feelers import Feelers
-from octomate.tentacles.lark.chromo import LarkChromo
-from octomate.tentacles.lark.feelers import (
+from octomate.tentacles.channel.base import ChannelTentacle, PlatformMessage
+from octomate.tentacles.channel.feelers import Feelers
+from octomate.tentacles.channel.lark.chromo import LarkChromo
+from octomate.tentacles.channel.lark.feelers import (
     LarkConfirmationFeeler,
     LarkQuestionFeeler,
     LarkTodoFeeler,
 )
-from octomate.tentacles.lark.ink import LarkInk
+from octomate.tentacles.channel.lark.ink import LarkInk
 from octomate.utils import guess_image_ext
 
 # Octopus imports tentacles (via connect_tentacles) so importing it here would
 # be circular at module load time.
 if TYPE_CHECKING:
-    from octomate.agents.pulse import PulseAgents
     from octomate.memory.base import OctopusMemory
     from octomate.octopus import Octopus
+    from octomate.tentacles.agent.pulse import PulseTentacle
 
 logger = logging.getLogger(__name__)
 
@@ -53,7 +53,7 @@ class LarkTentacle(ChannelTentacle):
         *,
         app_id: str,
         app_secret: SecretStr,
-        agents: PulseAgents,
+        pulse: PulseTentacle,
         memory: OctopusMemory,
         flush_delay: float = 0.5,
     ) -> None:
@@ -72,7 +72,7 @@ class LarkTentacle(ChannelTentacle):
             log_level=lark_oapi.LogLevel.INFO,
         )
         self.ws_scope = None
-        super().__init__(tag, octopus, agents, memory, flush_delay=flush_delay)
+        super().__init__(tag, octopus, pulse, memory, flush_delay=flush_delay)
 
         self.feelers = Feelers(
             confirm=LarkConfirmationFeeler(self.ink, self.interactions),
