+++ octomate/tentacles/channel/lark/feelers.py
@@ -8,7 +8,7 @@ import logging
 from typing import TYPE_CHECKING
 
 from octomate.stores.interaction import InteractionStore
-from octomate.tentacles.feelers import (
+from octomate.tentacles.channel.feelers import (
     ConfirmationFeeler,
     QuestionFeeler,
     QuestionResponse,
@@ -17,7 +17,7 @@ from octomate.tentacles.feelers import (
 
 if TYPE_CHECKING:
     from octomate.schemas.session import SessionKey
-    from octomate.tentacles.lark.ink import LarkInk
+    from octomate.tentacles.channel.lark.ink import LarkInk
     from octomate.transmuters.interactions import Confirmation, Question
 
 logger = logging.getLogger(__name__)
