+++ octomate/tentacles/channel/napcat/__init__.py
@@ -1,0 +1,8 @@
+from octomate.tentacles.channel.napcat.base import NapcatTentacle
+from octomate.tentacles.channel.napcat.ink import NapcatInk, NapcatUserProfile
+
+__all__ = [
+    "NapcatInk",
+    "NapcatTentacle",
+    "NapcatUserProfile",
+]
