+++ octomate/tentacles/channel/napcat/schema.py
