+++ octomate/tentacles/channel/slack/__init__.py
@@ -1,0 +1,8 @@
+from octomate.tentacles.channel.slack.base import SlackTentacle
+from octomate.tentacles.channel.slack.ink import SlackInk, SlackUserProfile
+
+__all__ = [
+    "SlackInk",
+    "SlackTentacle",
+    "SlackUserProfile",
+]
