+++ octomate/tentacles/channel/slack/ink.py
