--- octomate/tentacles/claude/__init__.py
@@ -1,3 +1,0 @@
-from octomate.tentacles.claude.agent import ClaudeCodeTentacle
-
-__all__ = ["ClaudeCodeTentacle"]
