--- octomate/tentacles/lark/__init__.py
@@ -1,8 +1,0 @@
-from octomate.tentacles.lark.base import LarkTentacle
-from octomate.tentacles.lark.ink import LarkInk, LarkUserProfile
-
-__all__ = [
-    "LarkInk",
-    "LarkTentacle",
-    "LarkUserProfile",
-]
