--- octomate/tentacles/napcat/__init__.py
@@ -1,8 +1,0 @@
-from octomate.tentacles.napcat.base import NapcatTentacle
-from octomate.tentacles.napcat.ink import NapcatInk, NapcatUserProfile
-
-__all__ = [
-    "NapcatInk",
-    "NapcatTentacle",
-    "NapcatUserProfile",
-]
