--- octomate/tentacles/slack/__init__.py
@@ -1,8 +1,0 @@
-from octomate.tentacles.slack.base import SlackTentacle
-from octomate.tentacles.slack.ink import SlackInk, SlackUserProfile
-
-__all__ = [
-    "SlackInk",
-    "SlackTentacle",
-    "SlackUserProfile",
-]
