+++ octotools/pixiv.py
@@ -13,7 +13,7 @@ from pydantic import BaseModel, SecretStr
 from pydantic_ai import RunContext
 from pydantic_settings import BaseSettings, SettingsConfigDict, YamlConfigSettingsSource
 
-from octomate.agents.manager import SkillManager
+from octomate.tentacles.agent.skills import SkillManager
 
 P = ParamSpec("P")
 R = TypeVar("R")
